installed react-bootstrap dependency, added login and signup css from separate feature branch, moved login and signup components to new folders within components. Updated background color in index.css. Refactored form code for login and signup components to match code on signup-and-login-components feature branch.
