+++ package.json
@@ -9,6 +9,7 @@
     "axios": "^0.21.1",
     "bootstrap": "^5.0.2",
     "react": "^17.0.2",
+    "react-bootstrap": "^2.0.0-beta.4",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
